Enemy Death & Health Functionality, Timer Fix on Main1, Initial Code for Alt Shoot, WandBullet Code for EnemyDamage
Enemies now have a healthpool and take damage after a EnemyDamage call on their script.  WandBullets have been set up to call this function upon a 2DArea collision signal that checks for enemies before calling said function.

Fixed Timer on Main1 that limited the duration to 1 minute because enemy spawns would stop.  We still need to fix the bug that causes crashes when the player can't find a target on-screen.

Initial code was written for altShoot in the MagicWand script.